A simpler version of the ExperimentalFrame - Experiment - Treatment - Replication framework has been implemented. For many purposes, this framework is way too complicated and rigid. It is now possible to quickly tell the simulator: run 1 replication of the current model; or run n replications of the current model. The order in which simulator and experiment need to be created is now totally free; when the  simulator.initialize(replication) method is called, both the simulator and the replication need to fully specified, of course. Note that the ReplicationMode has DISAPPEARED from the call to  simulator.initialize(...). It was not used and not stored, and belongs in the Experiment rather than in the call to initialize().
The Replication.START_REPLICATION_EVENT, Replication.END_REPLICATION_EVENT, and Replication.WARMUP_EVENT have been moved to ReplicationInterface.START_REPLICATION_EVENT, ReplicationInterface.END_REPLICATION_EVENT 

The ExperimentalFrame class has been removed from DSOL. Instead, a new Experiment class has been introduced.

+++ trunk/dsol-demo/src/main/java/nl/tudelft/simulation/dsol/tutorial/section42/Customer.java
@@ -49,7 +49,7 @@ public class Customer implements BuyerInterface
         this.retailer = retailer;
         StreamInterface stream = this.simulator.getReplication().getStream("default");
         this.intervalTime = new DistExponential(stream, 0.1);
-        EmpiricalDistribution empDist = EmpiricalDistribution.createFromDensities(new double[] {1, 2, 3, 4},
+        EmpiricalDistribution empDist = EmpiricalDistribution.createFromDensities(new long[] {1, 2, 3, 4},
                 new double[] {1.0 / 6.0, 1.0 / 3.0, 1.0 / 3.0, 1.0 / 6.0}, false);
         this.orderBatchSize = new DistDiscreteEmpirical(stream, empDist);
         this.createOrder();
